refactor(test): move calculateElapsedTime test to time.test.ts
